refactor: use OpenTripPlannerClient
- deprecate `TripPlan.Api` behaviour and the modules using it (`TripPlan.Api.MockPlanner` and `TripPlan.Api.OpenTripPlanner`) in favor of using the client (`OpenTripPlannerClient` and its behaviour `OpenTripPlannerClient.Behaviour`).
- remove `TripPlan.plan/4` function
