+++ config/dotcom/trip_planner.exs
@@ -1,11 +1,8 @@
 import Config
 
-config :dotcom, TripPlanApi, module: TripPlan.Api.OpenTripPlanner
-
 config :dotcom, TripPlanGeocode, module: TripPlan.Geocode
 
 if config_env() == :test do
-  config :dotcom, TripPlanApi, module: TripPlan.Api.MockPlanner
 
   config :dotcom, TripPlanGeocode, module: TripPlan.Geocode.MockGeocode
 end
