+++ config/runtime.exs
@@ -118,9 +118,9 @@ config :dotcom, OpenTripPlanner,
   wiremock_proxy: System.get_env("WIREMOCK_PROXY", "false"),
   wiremock_proxy_url: System.get_env("WIREMOCK_TRIP_PLAN_PROXY_URL")
 
-if config_env() != :test and System.get_env("OTP_URL", "") != "" do
+if config_env() != :test and System.get_env("OPEN_TRIP_PLANNER_URL") != "" do
   config :open_trip_planner_client,
-    otp_url: System.get_env("OTP_URL")
+    otp_url: System.get_env("OPEN_TRIP_PLANNER_URL")
 end
 
 if config_env() != :test do
