+++ test/trip_plan/trip_plan_test.exs
@@ -10,20 +10,4 @@ defmodule TripPlanTest do
       assert_received {:geocoded_address, "address", {:ok, ^position}}
     end
   end
-
-  describe "plan/4" do
-    test "returns {:ok, itineraries} from the api" do
-      {:ok, from} = geocode("from")
-      {:ok, to} = geocode("to")
-      connection_opts = [user_id: 1]
-      opts = [depart_at: ~N[2017-07-06T19:20:00]]
-      assert {:ok, itineraries} = plan(from, to, connection_opts, opts)
-      assert [%TripPlan.Itinerary{}] = itineraries
-
-      assert_received {:planned_trip, {^from, ^to, ^connection_opts, received_opts},
-                       {:ok, ^itineraries}}
-
-      assert received_opts[:depart_at] == opts[:depart_at]
-    end
-  end
 end
